services/freecad-worker/pyproject.toml
[project]
name = "cad-agent-worker"
version = "0.3.0"
requires-python = ">=3.10"
dependencies = [
  "redis==5.0.8",
  "rq==1.16.2",
  "boto3==1.34.162",
  "pydantic-settings==2.4.0",
  "httpx==0.27.2",
]

[tool.setuptools.packages.find]
include = ["worker*", "shared*"]
